feat(website): support hmr and react refresh
