+++ .gitignore
@@ -28,3 +28,4 @@ yarn-error.log*
 **/ssg
 
 types
+**/common-server
