+++ packages/website/babel.config.cjs
@@ -9,6 +9,7 @@ function isWebpack(caller) {
 export default (api) => {
   const web = api.caller(isWebTarget);
   const webpack = api.caller(isWebpack);
+
   const isDevelopment
     = !process.env.NODE_ENV || process.env.NODE_ENV === "development";
 
@@ -40,7 +41,7 @@ export default (api) => {
       ],
       ["@babel/plugin-syntax-dynamic-import"],
       ["@loadable/babel-plugin"],
-      isDevelopment && "react-refresh/babel",
+      isDevelopment && isWebTarget && "react-refresh/babel",
     ].filter(Boolean),
   };
 };
